Skip infoblox-client as automatic BR
infoblox-client is a requirement for optional driver. It is not in
Fedora so we can not import it in RDO. If someone is interested in this
driver, should add it to Fedora and then to RDO following the process to
add a new dependency.

In the meanwhile we will skip it as BR.

Change-Id: I0d498016793d24722dd12021b730822777d9d0e9